backend/pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "asgiref>=3.10.0",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi-mail>=1.5.0",
    "fastapi[all]>=0.115.14",
    "flower>=2.0.1",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "redis[hiredis]>=6.4.0",
    "scalar-fastapi>=1.0.3",
    "sqlmodel>=0.0.24",
    "twilio>=9.8.4",
]
